refactor(messages): refactor message feature in FastAPI, also moved websocket route to its own feature
