+++ backend/src/main.py
@@ -11,6 +11,7 @@ from .auth import public_router as auth_public_router
 from .messages import router as messages_router
 from .friends import router as friends_router
 from .notifications import router as notifications_router
+from .websocket import router as websocket_router
 
 
 def create_app():
@@ -37,6 +38,7 @@ def create_app():
     app.include_router(messages_router)
     app.include_router(friends_router)
     app.include_router(notifications_router)
+    app.include_router(websocket_router)
 
     return app
