+++ backend/src/websocket/router.py
@@ -1,0 +1,98 @@
+from fastapi import (
+    APIRouter,
+    WebSocket,
+    WebSocketDisconnect,
+    Depends,
+)
+
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from uuid import UUID
+
+import traceback
+from typing import Annotated
+
+from redis.asyncio import Redis
+
+from src.messages.exceptions import InvalidConversationID
+
+from src.core import manager, get_db, get_redis
+
+from src.auth.dependencies import get_current_user_id
+
+from src.auth.services import AuthService
+
+
+router = APIRouter(prefix="/api/websocket", tags=["Webscoket"])
+
+
+@router.websocket("/ws")
+async def websocket_endpoint(
+    websocket: WebSocket,
+    user_id: Annotated[UUID, Depends(get_current_user_id)],
+    redis: Redis = Depends(get_redis),
+    db: AsyncSession = Depends(get_db),
+):
+
+    await manager.connect(websocket, user_id)
+
+    # online:{user_id} is for general online state, while online_users is for group operations
+
+    # Only store "1" in redis, it is just a placeholder to know if user is online
+    await redis.set(f"online:{user_id}", "1", ex=60)
+
+    await redis.sadd("online_users", str(user_id))  # type: ignore
+
+    try:
+        online_direct_message_participants = await redis.sinter(f"user:direct_message:{user_id}", "online_users")  # type: ignore
+
+        for online_direct_message_participant in online_direct_message_participants:
+            online_status_dict = {"isOnline": True}
+
+            event_data = {"type": "USER_ONLINE", "data": online_status_dict}
+
+            await manager.send_to_user(
+                UUID(online_direct_message_participant), event_data
+            )
+
+        while True:
+            data = await websocket.receive_json()
+
+            if data["type"] == "PING":
+                await redis.set(f"online:{user_id}", "1", ex=60)
+                await redis.sadd("online_users", str(user_id))  # type: ignore
+                await websocket.send_json({"type": "PONG"})
+
+    except InvalidConversationID:
+        traceback.print_exc()
+        await websocket.close(code=3001)
+    except WebSocketDisconnect:
+        traceback.print_exc()
+        manager.disconnect(websocket, user_id)
+
+        remaining_connections = manager.get_num_of_current_connections_for_a_user(
+            user_id
+        )
+
+        if remaining_connections == 0:
+            online_direct_message_participants = await redis.sinter(f"user:direct_message:{user_id}", "online_users")  # type: ignore
+
+            await redis.delete(f"online:{user_id}")
+            await redis.srem("online_users", str(user_id))  # type: ignore
+
+            last_online = await AuthService.update_last_online(db, user_id)
+
+            for online_direct_message_participant in online_direct_message_participants:
+                offline_status_dict = {
+                    "isOnline": False,
+                    "lastOnline": last_online.isoformat() if last_online else None,
+                }
+
+                event_data = {"type": "USER_OFFLINE", "data": offline_status_dict}
+
+                await manager.send_to_user(
+                    UUID(online_direct_message_participant), event_data
+                )
+
+    except Exception:
+        traceback.print_exc()
